refactor: Usuario
