+++ FoodMinder/src/main/java/com/foodminder/FoodMinder/controllers/AuthenticationController.java
@@ -1,13 +1,10 @@
 package com.foodminder.FoodMinder.controllers;
 
-import com.foodminder.FoodMinder.domain.infra.security.TokenService;
 import com.foodminder.FoodMinder.domain.usuario.*;
+import com.foodminder.FoodMinder.services.AuthenticationService;
 import jakarta.validation.Valid;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.ResponseEntity;
-import org.springframework.security.authentication.AuthenticationManager;
-import org.springframework.security.authentication.UsernamePasswordAuthenticationToken;
-import org.springframework.security.crypto.bcrypt.BCryptPasswordEncoder;
 import org.springframework.web.bind.annotation.PostMapping;
 import org.springframework.web.bind.annotation.RequestBody;
 import org.springframework.web.bind.annotation.RequestMapping;
@@ -17,32 +14,14 @@ import org.springframework.web.bind.annotation.RestController;
 @RequestMapping("auth")
 public class AuthenticationController {
     @Autowired
-    private AuthenticationManager authenticationManager;
-    @Autowired
-    private UsuarioRepository repository;
-    @Autowired
-    private TokenService tokenService;
+    private AuthenticationService authenticationService;
 
     @PostMapping("/login")
     public ResponseEntity login(@RequestBody @Valid AuthenticationDTO data){
-        System.out.println("Tentando trocar usuario: " + data.login() + " e " + data.password());
-        var usernamePassword = new UsernamePasswordAuthenticationToken(data.login(), data.password());
-        System.out.println(usernamePassword);
-        var auth = this.authenticationManager.authenticate(usernamePassword);
-        System.out.println("Usuário autenticado com sucesso: " + data.login() + " e " + data.password());
-        var token = tokenService.generateToken((Usuario) auth.getPrincipal());
-        return ResponseEntity.ok(new LoginResponseDTO(token));
+        return authenticationService.login(data);
     }
-
     @PostMapping("/registrar")
-    public ResponseEntity register(@RequestBody @Valid RegisterDTO data){
-        if(this.repository.findByLogin(data.login()) != null) return ResponseEntity.badRequest().build();
-
-        String encryptedPassword = new BCryptPasswordEncoder().encode(data.password());
-        Usuario newUser = new Usuario(data.login(), encryptedPassword, data.role().toString().toUpperCase());
-
-        this.repository.save(newUser);
-
-        return ResponseEntity.ok().build();
+    public ResponseEntity registrar(@RequestBody @Valid RegisterDTO data){
+        return authenticationService.cadastrar(data);
     }
 }
