+++ FoodMinder/src/main/java/com/foodminder/FoodMinder/domain/usuario/AuthenticationDTO.java
@@ -1,4 +1,10 @@
 package com.foodminder.FoodMinder.domain.usuario;
 
 public record AuthenticationDTO(String login, String password) {
+    public String getLogin() {
+        return login;
+    }
+    public String getPassword() {
+        return password;
+    }
 }
