+++ FoodMinder/src/main/java/com/foodminder/FoodMinder/domain/usuario/LoginResponseDTO.java
@@ -1,4 +1,12 @@
 package com.foodminder.FoodMinder.domain.usuario;
 
 public record LoginResponseDTO(String token) {
+
+    public LoginResponseDTO(String token) {
+        this.token = token;
+    }
+
+    public String getToken() {
+        return token;
+    }
 }
