+++ FoodMinder/src/main/java/com/foodminder/FoodMinder/domain/usuario/Usuario.java
@@ -1,10 +1,7 @@
 package com.foodminder.FoodMinder.domain.usuario;
 
 import jakarta.persistence.*;
-import lombok.AllArgsConstructor;
-import lombok.EqualsAndHashCode;
-import lombok.Getter;
-import lombok.NoArgsConstructor;
+import lombok.*;
 import org.springframework.security.core.GrantedAuthority;
 import org.springframework.security.core.authority.SimpleGrantedAuthority;
 import org.springframework.security.core.userdetails.UserDetails;
@@ -12,6 +9,7 @@ import org.springframework.security.core.userdetails.UserDetails;
 import java.util.Collection;
 import java.util.List;
 
+@Builder
 @Table(name = "usuario")
 @Entity(name = "usuario")
 @Getter
@@ -32,10 +30,10 @@ public class Usuario implements UserDetails {
         this.role = role;
     }
 
-  @Override
-  public Collection<? extends GrantedAuthority> getAuthorities() {
-      return List.of(new SimpleGrantedAuthority("ROLE_" + role.toString()));
-  }
+    @Override
+    public Collection<? extends GrantedAuthority> getAuthorities() {
+        return List.of(new SimpleGrantedAuthority("ROLE_" + role.toString()));
+    }
 
     @Override
     public String getUsername() {
